I'm going to be turning everything into jsp
Things that are going to change:

NavBar is going to a jsp so we can simply include it in every file. I'll
go around and refactor everything.

Footer is also included, but it doesn't work i'm not sure why, someone
look into that.

web.xml welcome file is changed to index.jsp.


P.S. first time i load webpage Navbar is kinda wonky.